Deduplicate Azure CLI integration code
