+++ src/acs.ts
@@ -4,6 +4,7 @@ import { TextDocumentContentProvider, Uri, EventEmitter, Event, ProviderResult, 
 import { Shell } from './shell';
 import { FS } from './fs';
 import { Advanceable, Errorable, UIRequest, StageData, OperationState, OperationMap, advanceUri as wizardAdvanceUri, selectionChangedScript as wizardSelectionChangedScript, script, waitScript } from './wizard';
+import { Context, getSubscriptionList, loginAsync } from './azure';
 
 export const uriScheme : string = "acsconfigure";
 
@@ -20,11 +21,6 @@ enum OperationStage {
     PromptForSubscription,
     PromptForCluster,
     Complete,
-}
-
-interface Context {
-    readonly fs: FS;
-    readonly shell: Shell;
 }
 
 class UIProvider implements TextDocumentContentProvider, Advanceable {
@@ -106,24 +102,6 @@ function parseCluster(encoded: string) {
     return {
         resourceGroup: encoded.substr(0, delimiterPos),
         name: encoded.substr(delimiterPos + 1)
-    };
-}
-
-async function getSubscriptionList(context: Context) : Promise<StageData> {
-    // check for prerequisites
-    const prerequisiteErrors = await verifyPrerequisitesAsync(context);
-    if (prerequisiteErrors.length > 0) {
-        return {
-            actionDescription: 'checking prerequisites',
-            result: { succeeded: false, result: false, error: prerequisiteErrors }
-        };
-    }
-
-    // list subs
-    const subscriptions = await listSubscriptionsAsync(context);
-    return {
-        actionDescription: 'listing subscriptions',
-        result: subscriptions
     };
 }
 
@@ -399,47 +377,6 @@ function selectionChangedScript(operationId: string) : string {
 
 function advanceUri(operationId: string, requestData: string) : string {
     return wizardAdvanceUri(commandName, operationId, requestData);
-}
-
-async function verifyPrerequisitesAsync(context: Context) : Promise<string[]> {
-    const errors = new Array<string>();
-    
-    const sr = await context.shell.exec('az --help');
-    if (sr.code !== 0 || sr.stderr) {
-        errors.push('Azure CLI 2.0 not found - install Azure CLI 2.0 and log in');
-    }
-
-    prereqCheckSSHKeys(context, errors);
-
-    return errors;
-}
-
-function prereqCheckSSHKeys(context: Context, errors: Array<String>) {
-    const sshKeyFile = context.shell.combinePath(context.shell.home(), '.ssh/id_rsa');
-    if (!context.fs.existsSync(sshKeyFile)) {
-        errors.push('SSH keys not found - expected key file at ' + sshKeyFile);
-    }
-}
-
-async function listSubscriptionsAsync(context: Context) : Promise<Errorable<string[]>> {
-    const sr = await context.shell.exec("az account list --all --query [*].name -ojson");
-    
-    if (sr.code === 0 && !sr.stderr) {  // az account list returns exit code 0 even if not logged in
-        const accountNames : string[] = JSON.parse(sr.stdout);
-        return { succeeded: true, result: accountNames, error: [] };
-    } else {
-        return { succeeded: false, result: [], error: [sr.stderr] };
-    }
-}
-
-async function loginAsync(context: Context, subscription: string) : Promise<Errorable<void>> {
-    const sr = await context.shell.exec(`az account set --subscription "${subscription}"`);
-
-    if (sr.code === 0 && !sr.stderr) {
-        return { succeeded: true, result: null, error: [] };
-    } else {
-        return { succeeded: false, result: null, error: [sr.stderr] };
-    }
 }
 
 async function listClustersAsync(context: Context) : Promise<Errorable<string[]>> {
